requirements.txt
# Core dependencies
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0

# RAG & Vector Database
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1

# Groq API
groq==0.4.2

# Utilities
pydantic==2.5.3
httpx==0.26.0
numpy==1.26.3
tiktoken==0.5.2

# Async support
aiohttp==3.9.1
asyncio-throttle==1.0.2
